Added feedback for wrong input to the forms.

+++ src/redux/budget/budgetReducer.js
@@ -8,7 +8,7 @@ const budgetReducer = (state=budgetState,action)=>{
         case UPDATE_BUDGET:
             return {
                 ...state,
-                budget:parseFloat(action.payload)
+                budget:parseInt(action.payload)
             }
         default:
             return state
